requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp_socks==0.9.0
aiosignal==1.3.1
atomicwrites==1.4.1
attrs==24.2.0
cachetools==5.5.0
cffi==1.17.1
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
logging_notifications @ git+https://github.com/thecheetahcat/logging-notifications.git@master
matrix-nio==0.25.2
multidict==6.1.0
orderbook_handler @ git+https://github.com/thecheetahcat/orderbook-handler.git@master
peewee==3.17.6
pycparser==2.22
pycryptodome==3.21.0
python-dotenv==1.0.1
python-olm==3.2.16
python-socks==2.5.3
referencing==0.35.1
rpds-py==0.20.0
socket_client @ git+https://github.com/thecheetahcat/socket-client.git@master
sortedcontainers==2.4.0
unpaddedbase64==2.1.0
websockets==13.1
yarl==1.13.1
